requirements.txt
biopython>=1.81
pytest>=7.0
